Feature/stylesettings

+++ client/components/Settings/Services/ServiceItemData.jsx
@@ -1,11 +1,16 @@
 import React, {Component, PropTypes } from 'react';
 import { Map } from 'immutable';
-import {  Form, Field, IconButton, CardHeader } from '../../library';
+import {  Form, Field, Button, Header, Grid, Row, Col } from '../../library';
 import styles from './styles.scss';
 
-function isNumber(value){
-  return value && !/\D/.test(value) ? undefined : 'Please enter a number.';
-}
+const parseNum = value => value && parseInt(value);
+
+const maxLength = max => value =>
+  value && value.length > max ? `Must be ${max} characters or less` : undefined
+const maxLength25 = maxLength(25);
+
+const notNegative = value => value && value <= 0 ? 'Must be greater than 0' : undefined;
+
 
 class ServiceItemData extends Component {
   constructor(props) {
@@ -16,14 +21,22 @@ class ServiceItemData extends Component {
 
   updateService(values) {
     const { service } = this.props;
+    values.name = values.name.trim();
+
     const valuesMap = Map(values);
     const modifiedService = service.merge(valuesMap);
+
     this.props.onSubmit(modifiedService);
   }
 
   deleteService() {
     const { service } = this.props;
-    this.props.deleteService(service.get('id'));
+
+    let deleteService = confirm('Are you sure you want to delete this service?');
+
+    if (deleteService) {
+      this.props.deleteService(service.get('id'));
+    }
   }
 
   render() {
@@ -40,48 +53,56 @@ class ServiceItemData extends Component {
     };
 
     return (
-      <div className={styles.formContainer}>
-        <div className={styles.servicesForm}>
-          <Form
-            form={`${service.get('id')}Form`}
-            onSubmit={this.updateService}
-            initialValues={initialValues}
-          >
-            <div className={styles.servicesFormRow}>
-              <div className={styles.servicesFormField}>
-                <Field
-                  required
-                  name="name"
-                  label="Name"
-                />
-              </div>
-              <div className={styles.servicesFormField}>
-                <Field
-                  required
-                  name="duration"
-                  label="Duration"
-                  validate={[isNumber]}
-                />
-              </div>
-              <div className={styles.servicesFormField}>
-                <Field
-                  required
-                  name="bufferTime"
-                  label="Buffer Time"
-                  validate={[isNumber]}
-                />
-              </div>
-            </div>
-          </Form>
-        </div>
-        <div className={styles.trashButton}>
-          <IconButton
-            icon="trash-o"
-            className={styles.trashButton__trashIcon}
-            onClick={this.deleteService}
-          />
-        </div>
-      </div>
+      <Grid>
+        <Row className={styles.serviceHeaderContainer}>
+          <Header title={service.get('name')} />
+          <div className={styles.trashButton}>
+            <Button icon="trash" raised className={styles.trashButton__trashIcon} onClick={this.deleteService}>
+              Delete
+            </Button>
+          </div>
+        </Row>
+        <Row className={styles.formContainer}>
+          <Col xs={6} className={styles.servicesForm}>
+            <Form
+              form={`${service.get('id')}Form`}
+              onSubmit={this.updateService}
+              initialValues={initialValues}
+            >
+              <Row className={styles.servicesFormRow}>
+                <Col xs={12} className={styles.servicesFormField}>
+                  <Field
+                    required
+                    name="name"
+                    label="Name"
+                    validate={[maxLength25]}
+                  />
+                </Col>
+                <Col xs={12} className={styles.servicesFormField}>
+                  <Field
+                    required
+                    name="duration"
+                    label="Duration"
+                    type="number"
+                    normalize={parseNum}
+                    validate={[notNegative]}
+                  />
+                </Col>
+                <Col xs={12} className={styles.servicesFormField}>
+                  <Field
+                    required
+                    name="bufferTime"
+                    label="Buffer Time"
+                    type="number"
+                    normalize={parseNum}
+                    validate={[notNegative]}
+                  />
+                </Col>
+              </Row>
+            </Form>
+          </Col>
+        </Row>
+      </Grid>
     );
   }
 }
